refactor: extract settings classes
