+++ Editor/SmartBackups.cs
@@ -27,83 +27,6 @@ namespace AvatarSmartBackup
     }
 
     [Serializable]
-    public class BackupSettings
-    {
-        public bool autoRunOnLoad = true;
-        public int intervalMinutes = 10;
-        public bool debugMode = false;
-        public int keepSnapshots = 3;
-        public bool backupOnPlayEnter = true;
-
-        public bool incVRCAssets = true;
-        public bool incAnimControllers = true;
-        public bool incAnimationClips = true;
-        public bool incScenes = true;
-        public bool incMaterials = true;
-        public long materialsMaxKB = 1024;
-        public bool incDlls = true;
-        public long dllsMaxKB = 2048;
-
-        // UI presets for size limits (dropdown + custom)
-        // 0: 256 KB, 1: 512 KB, 2: 1024 KB, 3: 2048 KB, 4: 4096 KB, 5: Custom
-        public int materialsSizePresetIndex = 2;
-        public int dllSizePresetIndex = 3;
-
-        // Folders
-        public List<string> includeFolders = new List<string>() { "Assets/" };
-        public List<string> excludeFolders = new List<string>() { "Assets/StreamingAssets", "Packages" };
-
-        // ZIP / PERFORMANCE
-        public ZipPolicy zipPolicy = ZipPolicy.OnChange;
-        public int idleDelaySeconds = 10;             // Seconds of inactivity before zipping when Idle policy
-        public bool zipFastest = true;                 // Fastest vs Optimal
-        public bool autoThrottle = true;               // Automatic throttling (recommended)
-        public int copyMaxMBps = 250;                  // Manual cap (MB/s). 0 = unlimited
-        public int zipMaxMBps = 150;                   // Manual cap (MB/s). 0 = unlimited
-        public float lastMeasuredMBps = 0f;            // Result of last IO benchmark
-        public long lastBenchmarkTicks = 0;            // UTC ticks of last benchmark
-        public long lastBackupBytes = 0;                // Size of last backup data
-        public int maxParallelThreads = Math.Max(1, Environment.ProcessorCount);
-
-        public bool saveScenesBeforeBackup = false;    // Avoid blocking by default
-
-        public bool showAdvanced = false;
-        public bool useProjectSettings = false;
-
-        // Extension scoping: extensions (e.g., .prefab) from Folders & Types apply only within included folders when enabled
-        public bool extWithinIncludeFolders = true;
-    }
-
-    internal static class SessionKeys
-    {
-        public const string IsRunning = "ASB/IsRunning";
-        public const string NextRunUtc = "ASB/NextRunUtcTicks";
-        public const string LastBackupUtc = "ASB/LastBackupUtcTicks";
-        public const string RunsCount = "ASB/RunsCount";
-        public const string HookedVRC = "ASB/HookedVRC";
-    }
-
-    internal static class Session
-    {
-        public static bool IsRunning { get => SessionState.GetBool(SessionKeys.IsRunning, false); set => SessionState.SetBool(SessionKeys.IsRunning, value); }
-        public static DateTime? NextRunUtc { get => GetDT(SessionKeys.NextRunUtc); set => SetDT(SessionKeys.NextRunUtc, value); }
-        public static DateTime? LastBackupUtc { get => GetDT(SessionKeys.LastBackupUtc); set => SetDT(SessionKeys.LastBackupUtc, value); }
-        public static int RunsCount { get => SessionState.GetInt(SessionKeys.RunsCount, 0); set => SessionState.SetInt(SessionKeys.RunsCount, value); }
-        public static bool HookedVRC { get => SessionState.GetBool(SessionKeys.HookedVRC, false); set => SessionState.SetBool(SessionKeys.HookedVRC, value); }
-
-        static DateTime? GetDT(string k)
-        {
-            var s = SessionState.GetString(k, "");
-            if (string.IsNullOrEmpty(s)) return null;
-            if (long.TryParse(s, out var ticks) && ticks > 0) return new DateTime(ticks, DateTimeKind.Utc);
-            if (DateTime.TryParse(s, null, System.Globalization.DateTimeStyles.RoundtripKind, out var dt))
-                return dt.Kind == DateTimeKind.Utc ? dt : dt.ToUniversalTime();
-            return null;
-        }
-        static void SetDT(string k, DateTime? v) => SessionState.SetString(k, v.HasValue ? v.Value.ToString("o") : "");
-    }
-
-    [Serializable]
     public class ManifestEntry
     {
         public string guid;
